+++ neurodamus/node.py
@@ -2,6 +2,7 @@
 # Copyright 2018 - Blue Brain Project, EPFL
 
 from __future__ import absolute_import
+import gc
 import glob
 import itertools
 import logging
@@ -12,8 +13,6 @@ from os import path as ospath
 from collections import namedtuple, defaultdict
 from contextlib import contextmanager
 from shutil import copyfileobj, move
-import ctypes
-import ctypes.util
 
 from .core import MPI, mpi_no_errors, return_neuron_timings, run_only_rank0
 from .core import NeurodamusCore as Nd
@@ -33,41 +32,11 @@ from .neuromodulation_manager import NeuroModulationManager
 from .target_manager import TargetSpec, TargetManager
 from .utils import compat
 from .utils.logging import log_stage, log_verbose, log_all
+from .utils.memory import trim_memory, pool_shrink
 from .utils.timeit import TimerManager, timeit
 # Internal Plugins
 from . import ngv as _ngv  # NOQA
 from . import point_neuron as _point # NOQA
-
-
-def trim_memory():
-    """
-    malloc_trim - release free memory from the heap (back to the OS)
-
-    * We should only run malloc_trim if we are using the default glibc memory allocator.
-    When using a custom allocator such as jemalloc, this could cause unexpected behavior
-    including segfaults.
-
-    * The malloc_trim function returns 1 if memory was actually
-    released back to the system, or 0 if it was not possible to
-    release any memory.
-    """
-
-    if os.getenv('LD_PRELOAD'):
-        if 'jemalloc' in os.getenv('LD_PRELOAD'):
-            logging.warning("malloc_trim works only with the default glibc memory allocator. "
-                            "Please avoid using jemalloc (for now, we skip malloc_trim).")
-            logging.info("malloc_trim: not possible to release any memory.")
-            return 0
-
-    try:
-        path_libc = ctypes.util.find_library("c")
-        libc = ctypes.CDLL(path_libc)
-        logging.info("malloc_trim: memory released back to the system.")
-        return libc.malloc_trim(0)
-    except OSError:
-        logging.error('Unable to call malloc_trim.')
-        logging.info("malloc_trim: not possible to release any memory.")
-        return 0
 
 
 class METypeEngine(EngineBase):
@@ -1433,6 +1402,13 @@ class Node:
 
         Node.__init__(self, None, None)  # Reset vars
 
+        # Shrink ArrayPools holding mechanism's data in NEURON
+        pool_shrink()
+
+        # Garbage collect all Python objects without references
+        gc.collect()
+
+        # Finally call malloc_trim to return all the freed pages back to the OS
         trim_memory()
 
     # -------------------------------------------------------------------------